Start a new module for benchmark. Implement SorterStudy that replace GroundTruthStudy
