+++ src/spikeinterface/benchmark/benchmark_plot_tools.py
@@ -1,0 +1,9 @@
+
+
+
+
+def _simpleaxis(ax):
+    ax.spines["top"].set_visible(False)
+    ax.spines["right"].set_visible(False)
+    ax.get_xaxis().tick_bottom()
+    ax.get_yaxis().tick_left()
